backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
pytest
python-dotenv
python-jose
